+++ src/Modules/apiConsts.js
@@ -1,0 +1,8 @@
+export const url = 'https://shazam.p.rapidapi.com/charts/track?locale=en-US&pageSize=20&startFrom=0';
+export const options = {
+  method: 'GET',
+  headers: {
+    'X-RapidAPI-Key': '36333753fcmsh2c569f6a18ea132p116dc2jsn138253265de7',
+    'x-rapidapi-host': 'shazam.p.rapidapi.com',
+  },
+};
